astro-engine/requirements.txt
pyswisseph==2.10.3.2
fastapi==0.109.0
uvicorn==0.27.0
python-dateutil==2.8.2
timezonefinder==5.2.0
pytz==2024.1
geopy==2.4.1
